Update patch set 1
Patch Set 1:

If you look more closely at the link, it says:

> must follow the ISO 8601 standard: YYYY-MM-DD, shortened to just YYYY-MM or simply YYYY.

Since, as you say, Android's Media stuff doesn't do months, I think we can just do years and be done with it. Whether or not to extract YYYY from a full ISO 8601 could be an extra.

And right now, since neither Android/Java nor Cyanogenmod seems to have any reliable code to do that, I'm inclined to just check that it's a four-digit string and use that as a year. So basically, your second option.

In most languages I would do this via a very simple regex, something like "^[0-9][0-9][0-9][0-9]$". Would you do it this way?

Patch-set: 1