Update patch set 1
Patch Set 1:

And regarding the whole "ID3 vs VorbisComment" stuff - This is not a patch to add VorbisComment support. For all I know, the basics in VorbisComment work right now (artist, title, album, track) and it's only year that's bothering me.

This is just a way to add broader support to the parser. In order to add VorbisComment support, there would probably need to be a lot more work. I just care about a few basic media tags, so that is not my intention :)

Patch-set: 1