Update patch set 3
Patch Set 3:

Yeah... the thing is, I tried to properly build cyanogenmod for my nexus 6 to test this, but after hours on my slow connection and taking up way too much space on my SSD, I opted for just cloning this repo and blindly coding on it.

So it probably builds and works, but it might not. If it doesn't, apologies. This approach was fine when my patch was one line, but it's getting a bit out of control :)

Patch-set: 3