Update patch set 4
Patch Set 4:

> Thanks for sharing that. This really helps me to test. I'm actually
 > finding that even with this patch, I'm not seeing Year being
 > displayed in Vanilla Music. Is it working for you?

As mentioned above, I currently can't build images so I can't test it. Have you triggered a media storage rescan in some way?

I can have a closer look at the file later to confirm that it indeed has a DATE tag. Then a few prints could be added to the java code to see if the tag makes it to the code introduced by the patch.

Otherwise, if you build an image for shamu (nexus 6) I'd be able to test it.

Patch-set: 4