Update patch set 4
Patch Set 4: Code-Review-1

This is not enough to use the "date" tag. We need to add the tag to https://github.com/CyanogenMod/android_frameworks_av/blob/cm-12.0/media/libstagefright/StagefrightMediaScanner.cpp#L178 in order to properly mapping the tag and allow parse it in MediaScanner. Otherwise the tag never reach the handleStringTag method.

I will try to add the needed c++ code to made this patch useful.

Patch-set: 4