Update patch set 1
Patch Set 1:

IMHO icon - icon - text - text is a bit unexpected though, and icon - text - icon - text would be more intuitive.

If it's impossible to do this, I think I'd rather prefix the texts with B and D instead of underlining it, as the underline doesn't have any obvious relation to the dock.

Patch-set: 1