Update patch set 1
Patch Set 1:

" icon - text - icon - text" was my first try, but it causes undesired effects (reverted in http://review.cyanogenmod.org/#/c/103778). I don't know how fix it if text views aren't part of system_icons.xml layout. Any idea is welcomed :)

About the underlying thing, this patch is for a few devices. It's only a way to differentiate it (B and D doesn't mean to much if they are not explained).

Patch-set: 1