Update patch set 4
Patch Set 4:

Jorge,
It's an app that I have been working on internally that uses the shared user ID of system.  It publishes some custom tiles and when I would long press on one of those tiles it would crash SystemUI because it made the assumption that we were a dynamic tile, which is not the case.

Patch-set: 4
Reviewer: Clark Scheff <9199@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>