Update patch set 2
Patch Set 2:

> Is this a confirmed working fix? In a cm-12.1 build from today
 > (which still has this patch: https://github.com/CyanogenMod/android_frameworks_base/commits/cm-12.1/services/core/java/com/android/server/AssetAtlasService.java
 > ) I just hit this in my logs: http://paste.ubuntu.com/12010543/
 > 
 > Was attempting to encrypt a device; unsure if this crash is at all
 > involved in the failure to finish encryption.

This change checks if it is empty, so it does not try to access an item inside an empty list.

In your example the comparator is trying to access an index, which is not available, this code path is accessed asynchronously (we are creating multiple workers above) and needs some synchronization if i am correct.

Patch-set: 2