Update patch set 1
Patch Set 1: -Code-Review

Not a bad thing actually but is there any real usability of this for example you buttons are the problem here but in reality it your case which is a problem. You can change your case why would anyone would like to wake the device to just change the volume of the the playing media. You can use the volume buttons for this.
And also this makes it more gimmicky sort of a thing. It like hey there are other tiles commit let's see if I can make one of my own and then these unusable thing get into the aio concept. If you really want to do things as good think of finger gestures like in Oppo's color os. 

There are so mant tutorials to libraries available to refer to for those things.

Even see the cmsdk part of this cm hasn't taken away the possibility to add stuffs Om your own.
Just don't do the easy way around .

I am not against customization. But now there are too many tiles available by default in cm adding more and more stuffs makes it cumbersome and awkward . just think the OEM way of doing stuffs not just the user way. And as a developer the first principles :- Always concentrate on the user and usability part of things

Peace!

Patch-set: 1