Update patch set 2
Patch Set 2:

Yes, http://review.cyanogenmod.org/#/c/105883/ fixes the issue where people could not decrpyt their device because the keyguard sim unlock was being shown too early.  With the above mentioned patch, the sim unlock will not be shown until the device is decrypted and the system is up and running.

Patch-set: 2