Update patch set 1
Patch Set 1:

> IHMO, hardware volume keys are there for something. There isn't
 > need to have duplicate functionality.

I kind of agree with you on the point that there is no need for duplicate functionality. 
But for example, the double tap to sleep on the status bar is also not really necessary, we already have the power button.
IMHO this slider is really handy on big phones, fat cases and tablets when watching a video or listening to music with the screen on.
Also, it's turned off by default! :)

Patch-set: 1