Update patch set 10
Patch Set 10:

I don't think we need to migrate the setting because current values are 0 and 3 (BATTERY_PLUGGED_AC|BATTERY_PLUGGED_USB). And in case we need to migrate it, then it should be done in the DatabaseHelper of the settings provider.

I'm going to submit a new patch without the migration. Please feel free to revert it.

Patch-set: 10