Update patch set 1
Patch Set 1: Code-Review+1

Thank you, but your sugestions don't work. I extensively debugged this. Line 740 is correct because if you debug the code you will note that the notifications are comming from NetdResponseCode.InterfaceClassActivity. If you switch the fromRadio parameter to false the bug will trigger again. The lines 1515 and 1544 are irrelevants (seems to never being run).

If you look further in the code you will also note that the interface is already tested for type in the line:
final boolean isMobile = ConnectivityManager.isNetworkTypeMobile(type);

So your expression "ConnectivityManager.TYPE_MOBILE == type" is irrelevant and redundant. In fact the fromRadio parameter apears to serve to no real purpose at all in the notifyInterfaceClassActivity method.

Anyway, people are already using this patch and it fix the issue nicely (unreal high mobile radio active time). There is also an xposed module that implements this patch.

Patch-set: 1
Label: Code-Review=+1