Update patch set 1
Patch Set 1: Code-Review-1

This is not the correct solution to the bug.  This simply masks the reporting to make it look normal, when in fact there still is an underlying issue.  In the constructor for NetworkManagementService, there is a block that handles getting data connection information from the radio.  Here is that block:

mPhoneStateListener = new PhoneStateListener(SubscriptionManager.DEFAULT_SUBSCRIPTION_ID,
         mDaemonHandler.getLooper()) {
     @Override
     public void onDataConnectionRealTimeInfoChanged(
            DataConnectionRealTimeInfo dcRtInfo) {
        if (DBG) Slog.d(TAG, "onDataConnectionRealTimeInfoChanged: " + dcRtInfo);
        notifyInterfaceClassActivity(ConnectivityManager.TYPE_MOBILE,
                dcRtInfo.getDcPowerState(), dcRtInfo.getTime(), true);
    }
};

This is the ONLY time the fromRadio argument should be true because this callback triggers directly from radio reports.  All other calls to notifyInterfaceClassActivity do not come directly from radio information, and, as such, the fromRadio argument is false.  In addition, cursory examination of the logic flow in notifyInterfaceClassActivity reveals that when the trigger is from mobile networks and fromRadio is false, it uses the last reported power state from the radio to update powerState, which is then used to determine if the radio is in the medium or high DC power state, which is considered active.  This indicates that the mobile radio active issue is occurring somewhere deeper in the system where the radio is controlled.

Also, to dispel another assertion, callbacks received from NetworkManagementService do not have any impact on whether the radio is active or not.  NetworkManagementService exists to handle high level operations for enabling/disabling networks, handling firewall setup, and reporting the status of those networks.

Patch-set: 1
Label: Code-Review=-1