Update patch set 1
Patch Set 1:

That is the EXACTLY block that inserted the bug. If you look in the log history of the file you will see that. Also look at the kitkat or older code and you will see that the netd notifications will allways make to the observers notifications.

Before that block, the network notifications, network event observers and battery stats are working properly.

It looks like I am the only one here debuging the code. In this case the notifications are comming from netd and nonwhere else. The guy that created that block made some mistake and the block never sees the light of day.

I will state this for one more time: the fix will make the observers nofications and battery stats receiving the correct notification (debug the code to see that). This forum are full of devs merging code without proper review from others, so why not repopick this and try it yourself?

I'm starting to feel uncomfortable for posting this fix on this forum. It's funy. It's so easy to come here and stare 1 minute at code and do some commenting and conjecturing.

Patch-set: 1