Update patch set 1
Patch Set 1:

"Battery stats is notified of radio changes in a different way."
What way? You are not clear.

"There is a direct call noteMobileRadioPowerState in notifyInterfaceClassActivity that properly updates battery stats on mobile radio power changes."
This call only runs when fromRadio=true (the patch makes fromRadio=true)

"Not notifying the observers is by design from what I can tell."
Wrong. Based on what assumption? The code says the contrary of it.

But please enlighten me what you want at code level? Whats is missing to solve this issue? The observers and battery stats are being notified at data connection and disconnection. Please tell what is missing and I will look for?

I am on this for at least a week, and I am hot on the code.

Patch-set: 1