Update patch set 1
Patch Set 1:

"This call only runs when fromRadio=true (the patch makes fromRadio=true)"

Nope, check again.  Battery stats is notified when isMobile is true, which is determined by connection type.

Just because the fix appears to fix the issue doesn't mean it's the correct fix.  In this case it is incorrect to make fromRadio true 100% of the time because it forces the observers to be notified when they shouldn't.

Patch-set: 1