Update patch set 1
Patch Set 1:

Right off the bat you're wrong. 

"The line 253 generetas a one time notification with fromRadio=true"

Line 253 in CM (242-246 in AOSP) simply sets up the phone state listener.  It doesn't actually do anything until it receives the necessary callback.

"BUG because fromRadio=false then current state=mLastPowerStateFromRadio (current state discarded - PROBLEM: it was inactive, but is loaded from last, wich is active)"

Not a bug.  The radio is treated as authoritative for all power changes, so the power level will not be updated if the radio is not providing the information. When the radio is providing power information, battery stats is properly notified without the use of the observer (see 422 AOSP).

Patch-set: 1