Update patch set 1
Patch Set 1:

"The line 253 generetas a one time notification with fromRadio=true"
"Line 253 in CM (242-246 in AOSP) simply sets up the phone state listener.  It doesn't actually do anything until it receives the necessary callback."

This is not what the logs are saying. There is comming ONE and ONLY ONE (EVER) notification when the listener are created.


"Not a bug.  The radio is treated as authoritative for all power changes, so the power level will not be updated if the radio is not providing the information. When the radio is providing power information, battery stats is properly notified without the use of the observer (see 422 AOSP)."

This is not what the logs and the code are saying. There is no such authorithy. Tell me what authority is generating the notifications? Or should be? In the code?
So I am looking for a bug based on the the code and the logs. I can't solve inexistent things. If someone in the past was planning to create such a radio authority (for what purpose?) and forgot to update the code, then I'm sorry.

The logs and the code shows that the authority here is NetdResponseCode.InterfaceClassActivity gererating inactive/active notications.

But I insist. What part of the code including the side effects is wrong? The observers are being notified when they should, and not being notified when they should'nt. Please show me a case when this does not works and I will fix the patch.

Patch-set: 1