Update patch set 1
Patch Set 1:

"This is not what the logs are saying. There is comming ONE and ONLY ONE (EVER) notification when the listener are created."

If this is true then there is a bug somewhere else and it needs to be fixed, not masked by simply forcing updates.

"This is not what the logs and the code are saying. There is no such authorithy. Tell me what authority is generating the notifications? Or should be? In the code?"

Read the comments made by google engineers.

Lines 432-435
"// Report the change in data activity.  We don't do this if this is a change
// on the mobile network, that is not coming from the radio itself, and we
// have previously seen change reports from the radio.  In that case only
// the radio is the authority for the current state."

Patch-set: 1