Update patch set 1
Patch Set 1:

"-I, a volunteer, do the work of the Google Enginners and fix their bugs and construct the missing radio authority notification module"

I... what??? There is no code related to this... I was speaking in terms of how the data is treated.  Power level reports from the radio are the only reports that can change the reported active state.  All other states are ignored.

Patch-set: 1