Update patch set 1
Patch Set 1:

Ok.
The NetdResponseCode.InterfaceClassActivity is the currently radio power notification engine (it was on Kitkat).
The google enginners are (maybe) planning to switch, but there is no other authority to come from the notifications (not in the code).
The enginners simply forget to mark NetdResponseCode.InterfaceClassActivity notification as fromRadio=true in the code (in the antecipation of the new "missing" engine).
fromRadio=true are not masking anything.

Patch-set: 1