Update patch set 1
Patch Set 1:

Ok.
And then hows about people on Lollipop? Backporting the code?
And then hows about people on stock?
I will keep debugging on this. hehehe.
It's funny to me. I'm learning more about the Android internals. Maybe I become a frequent contributor.
Anyway people are already happy with the patch. Soon people will start to switch ROM to that with the patch. I'm happy with that tooo.

Patch-set: 1