Update patch set 1
Patch Set 1:

Yes. The issue is "cell standby mobile radio active issue" (like the title of this topic).

This patch fix exactly that. Nothing more, nothing less.
The battery stats and observers are being notified as expected (which wasnt with the bug), now resulting in correct mobile radio timing on battery stats.

I don't understand what are you expecting from a fix for "the cell standby mobile radio active issue" that fix the "cell standby mobile radio active issue" clean and without side-effects.

The bright new PhoneStateListener class implemented in the patch 2d4be (referenced on this topic) never worked properly as you can see. So this patch semanticaly reverts the mistake done on the two failed commits. So insted of doing googles job this is a simply fix to that mistake.

Patch-set: 1