Update patch set 1
Patch Set 1:

Woah you guys are flooding my inbox. Well, it is okay if the patch fails to meet the requirement of review. changing 3 variables arises suspicion that the google software engineers are wrong. The fact they are right and I trust the google engineers who did this original patch than xda itself because they are like 500x time smarter than the average xda user. Unlike the average xda users, they have a degree and/or years of writing software. Which is the majority of cyanogen and cyanogenmod community members have. Bottom line, this is a bad patch which should not be merged at all.

Patch-set: 1