Update patch set 1
Patch Set 1:

Come on, just merge and release the patch already. If it works, it will work. If it won't, just revert it, as simple as 1+1=2. When somebody trying to debug and y'all are trying to discredit. Bunch of disgraceful brats.

Patch-set: 1
Reviewer: Jefferson Cheang <13990@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0