Update patch set 1
Patch Set 1: Code-Review-1

No one is being disgraceful. 

This patch is essentially applying an incorrectly sized bandaid on an issue that seems to have been exhaustively debugged, and it may be causing you to believe what you are inferring from the logic flow. 

Unfortunately, there is a bigger/different root issue here and it would be silly/<some other word for a bad idea> to merge this patch as it DOESNT fix the issue.

 As Danny noted, if the callback is only ever called once then you need to trace it back to its origin, as that will provide a better picture as to what the actual issue is.

Patch-set: 1
Label: Code-Review=-1