Update patch set 1
Patch Set 1:

In particular, the onDataConnectionRealTimeInfoChanged callback should be called from DcController.java:374, which is in turn fed from the isAnyDataCallActive variable in that method, which is fed from the active member of the DataCallResponse object coming from the (device specific) RIL.

If that callback is called only once, it's probably worth investigating what the exact values of that active member are. It's not unlikely that the RIL doesn't report this correctly.

Patch-set: 1