Update patch set 1
Patch Set 1:

I agree with the changes, however you have to remember that a user can set the default color at a very low brightness, whereas the applications with custom color can use a 100% bright color. So the selector was meant to display a full white in order to represent the real brightness master control.

Patch-set: 1