Update patch set 1
Patch Set 1:

At first I used an hardcoded 0xFFFFFF because I thought the same thing, but then I decided to use the default color. That's how users likely want their LED to be and so using it for the preview is not a bad idea.

It's also true that each application uses the color it wants, so a neutral white might be better as it gives a better idea of the brightness level.

The important thing for me here is not to drop the brightness value.

Patch-set: 1