Update patch set 4
Patch Set 4:

Yeah, NfcService register itself in the ServiceManager, but since it is a separate process, it can end a race condition with systemui startup, ending in that the service is not ready yet when nfctile were created.

We need to teawk Nfc app and SystemServer to start it during booting process. I will try to add the pieces to start the service earlier than systemui.

Patch-set: 4