Update patch set 1
Patch Set 1: Code-Review-1

Jorge: yes, it does. I'm not sure that's a bad thing, tho, nor that the intended behavior in the case where one of the volumes is null is to stop in that one, and instead mark the primary physical volume as MEDIA_REMOVED. 

In summary: you're right, this needs to be better studied, and we need to know what exactly are the repercussions of this change (or if we need to do more changes paired with it).

Patch-set: 1
Label: Code-Review=-1