Update patch set 1
Patch Set 1:

My concerns about set the visualizer to null is that at that point of the code visualizer always will be null, because if you call link and get a valid ref, then you only call link again if you previously call unlink which set the ref to null. So there is no reason to set if again to null.

About the listener, 2 lines below you call to release that should remove the listener and any capture process. That's why i don't understand this path. Data from listener should stop even without set the listener to false. If that's not true, then the fix should be done in the native code and not here.

Patch-set: 1