Update patch set 2
Patch Set 2:

Maybe it should, but since the documentation _explicitly_ mentions it should be set to null I don't want to mess with public APIs.

The release method does call native_release() but does not call native_setPeriodicCapture(). Again, I'm not sure if I want to mess with the public APIs of how things are expected to work.

Patch-set: 2