Update patch set 2
Patch Set 2:

Hm not sure I understand your concerns, when a process picks up focus for the mic, we are using the process's pid to determine which app (package name) this focus corresponds to. 

This is just a more specific way of determining which package has mic focus compared to UID which can be shared amongst packages in separate processes.

Patch-set: 2