Update patch set 2
Patch Set 2:

This code gets executed literally right as a process requests audio recording or stops it. It seems highly unlikely that process dies as this happens. And even if it does, it'd be appropriate to return null here.

I think using PID to map to a package name is far more specific than using the UID in this case. And keep in mind we're not storing the PID, we're only persisting the package name. This isn't meant to persist across reboots or the like. This PID gets used to find the package, and keep that in the process memory while the device is up. It's meant to get reset on reboot or the process dying.

Patch-set: 2
Reviewer: Rajesh Yengisetty <9528@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0