Update patch set 1
Patch Set 1:

IMHO, it's better to add a static getInstance(Context) method in RecentsConfiguration and check whether sInstance is null, and in case create a new reference of it, because RecentsTaskLoadPlan(context) has a valid context to create the instance. This way you be sure that mConfig never will be null in RecentsTaskLoadPlan.

Patch-set: 1