Update patch set 1
Patch Set 1:

I think is not a good idea, because this allow third party apps to bypass all the legal considerations about capture audio without inform the user. Amend, an app can grab audio without the consent of the user. IMHO, it should be restricted to system, but it's just my opinion.

Patch-set: 1
Label: Code-Review=0