Update patch set 1
Patch Set 1: Code-Review-1

Like Hexchain mentions, this can be done by configuring the setting, but we should do it as a tiered fallback.

We can change the logic for captive portal server by placing the value for it within the resources for values-mcc460 and querying that value first, if the value is null or empty, then we fallback to the default Settings.Global.CAPTIVE_PORTAL_SERVER, and if that's empty or null, default back to the default server.

Patch-set: 1
Reviewer: Adnan Begovic <9813@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1