Update patch set 2
Patch Set 2:

I just pushed up a new version of the healthd patch in this series ([2/3]) to make this extensible.

This was nice because I could revert my kernel changes to map our kernel's sysfs support in one that matched my moto kernel.

An example of using the libhealthd for this very non-standard case is:

https://github.com/CM-zenfone2/android_device_asus_mofd-common/commit/a05e3455d668f8bd6800279eab2bb4e537a66b11

(but you need to pretend I didn't accidentally include GCC := 4.9 in there, that was a mistake that I fixed).

This should allow us to support any device that has a sysfs prop that identifies whether or not it is fast charging.

Patch-set: 2