Update patch set 4
Patch Set 4:

> (1 comment)

That line is the AOSP behaviour. In AOSP (or before this patch) there is no secondary click. A, let's say, normal click will give you either the details (if device supports) or the DATA_USAGE_SETTINGS, if not.
I added that same behaviour on secondaryClick to be like wifi and bluetooth's tiles. The normal click is just for toggling data, but it still has the AOSP behaviour if your device do not have data.

If this doesn't make sense maybe I'm not understanding your point.

About the secondaryClick on dual qs tiles: here, http://review.cyanogenmod.org/#/c/110087/, I'm suggesting an uniform way of handle this and add the abbility to use secondaryClick's on non-dual tiles. By now, each tile has its own behaviour which is odd and confusing.

Patch-set: 4