Update patch set 1
Patch Set 1:

Isn't this too costly to use for every notification? There was some improvements around reduce the notification display times, and i think this will increase them again. The idea is really good and this code should be move to a helper class somewhere, but i think for improve the code we could so a neutral color (black) or inactive color and generate the palette in async mode (activating the led once the color for the app was generated)

Patch-set: 1