Update patch set 1
Patch Set 1:

I'm not a fan of making system|signature items available to data/apps, as that breaks the permission model to a degree. 

That said, I could see a new change with a custom permission that allows for this to be introduced - this would prevent any potential implications to 3rd party apps that may declare this in their manifest erroneously.

Patch-set: 1