Update patch set 3
Patch Set 3:

I added a "+" to the portait and landscape style battery icon (as opposed to the circle).

This seems to be working for me on victara (which was crashing when Tortel tried to test it).  However, I don't actually have a qcom fast charger right now so my testing is all faked at the moment.

Patch-set: 3