Update patch set 6
Patch Set 6:

No reason that it wouldn't do wouldn't with hdvcp.  Victara is using it and it works for it.

The only dependency is that there is a /sys indicator of whether or not it is fast charging.

Of course I was just told that Google is supporting this in M so that might be the kiss of death for this patch series.

Patch-set: 6