Update patch set 4
Patch Set 4:

state.visible = !mPm.hasPowerProfiles();

This made the tile not visible if the devices doesn't have profiles, or i missing something.
Also, PerformanceManager#setPowerProfile will throw an IllegalArgumentException if power profiles aren't enabled. What I'm trying to arguee is that PowerManager#setPowerSaveMode should doesn't depend on call to setPowerProfile because some devices couldn't implement it. Is that right or i'm totally wrong. setPowerProfile is called many times without checking, so maybe i'm wrong and it's safe to call it. In that case, this patch looks good then.

Patch-set: 4