Update patch set 1
Patch Set 1: Code-Review+1

Actually I would like to have this. Is that duplicated? Yes... But CM also have a duplicated option to end a call. A threeplicate-buggy way to see received SMS, a duplicate way of wake up device, etc. If one doesn't like this because duplication... The he might be using the wrong projects.

Patch-set: 1
Label: Code-Review=+1