Update patch set 1
Patch Set 1:

I agree. But this only kicks in if the percentage is zero. And in that scenario, it's safer to get false positives and kick the device to charger mode than it is to let it keep pulling current. As soon as it reaches 1%, this check becomes moot.

Patch-set: 1