Update patch set 6
Patch Set 6:

(5 comments)

Danny,

Thanks for the feedback.  I tried to address all of them.

However, I'm not really sure that we're ever going to merge this as I understand that Google is adding this feature in M and I was just a little too late in getting it added in CM before they stole my thunder.

If Google's changes are't intertwined with too much else, we probably want to backport that commit (or commits) instead.  Although I bet we'll probably need the healthd commit to let us have more flexibility in how it works.

Patch-set: 6