Update patch set 2
Patch Set 2:

> Yeah, I don't like this at all, whats the use case even? 
Follow the link in the commit description. It clearly shows a use case. This is also not theoretical: [1]

> Just open up a gaping hole in the security of the OS.
This patch is actually no influence on the security of the OS, maybe on badly-written third party applications. As mentioned in the link, this happens on "app-level" and the package manager does not use the spoofed signature.

[1] http://forum.xda-developers.com/android/apps-games/app-microg-gmscore-floss-play-services-t3217616

Patch-set: 2
Reviewer: la|r|ma <13914@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0