Update patch set 2
Patch Set 2:

> "Maybe on badly written apps" already disqualifies this. You should
 > be able to write the worst app known to man and not break the
 > security model.

I guess you misunderstood. It has influence on the application, but not in terms of security. The application has the same permission as before. If it was using permissions that with protectionLevel signature, still only those with the same signature can gain them. If it was using sharedUid, still only packages with the same signature can share the uid.

As long as Security is not inferred from random facts, this is no problem. Of course a DRM system that works by checking for a certain package to have a certain signature can be broken this way, but a DRM system that checks for a certain file to be existing on the file system can also be broken easily. And the mimimal possible influence that could lead to problems on badly designed applications is the reason to protect this feature with a "dangerous" permission.

Patch-set: 2