Update patch set 2
Patch Set 2:

The idea is to allow third-party apps to use it.

Of course mimicking play services will allow applications to get "all the data" that is send to play services from the phone (usually that's not to much, when you think about it). And note that you can only have one Play services installed.

Sensitive data should normally be protected by permissions.

The "dangerous" permission is just to ensure, that the user knows about the possible impact the app can have before installing. Your argumentation does not make sense, considering that a lot more data can be reached by asking the user for root permission.

Finally, yes, the idea of the patch is to punch a "hole" into the "security model" (mostly obscurity) inside play services. If Google would have designed it in a way, that does not force the implementation of such a patch, I would be grateful, but that's not the case.
As a "lower-impact" version I could imagine to restrict the feature to certain package names, but following your current points I guess this would not make you happy.

So I think we can stop here and accept that CM users will continue to punch a real "hole" in the security system of android by using Xposed instead to provide the same feature.

(People told me that CM is unlikely to accept the patch, but I did not want to belive and apparently the rumors are true and CM is no longer the ROM "by the community, for the community".)

Patch-set: 2