Update patch set 7
Patch Set 7:

After some mockups and discussions with Asher, I think we've come to the conclusion that the battery icon in the status bar is not the best way to handle the extra information. However, we do have dynamic tiles in the quick settings which would be a good solution for this scenario.

When fast charging, we could show a tile with an icon and label to indicate the status. The label could be "Fast Charging" - or something similar - and we can craft an icon for this. If you press the tile, it could just take you to the battery settings. 

Thoughts?

Patch-set: 7
Label: Code-Review=0