Update patch set 1
Patch Set 1:

So, I know you wanted to use the material icon for image, but Asher and I thought it might be cool to do something unique for CM. I've mocked that idea up.

http://i.imgur.com/NTJmoUR.jpg

I also created the error state for the icon you had chosen. That's included in the mock as well.

Thoughts?

Patch-set: 1