Update patch set 3
Patch Set 3:

Roman, what about retain the ArrayMap structure and instead of use an iterator just do a reverse iterate using keyAt and valueAt?. This way you don't need a synchonized block (remove items are safe, and add goes to the end of the array, this time it doesn't get a notification).
In that case the synchonized collections should be enough, though.

Patch-set: 3