Update patch set 5
Patch Set 5:

In PS4, couldn't we easily get into a state where N is 5; then a client is removed (mClients size is now 4) and when we try to access mClients.get(N==5) we'd get an out of bounds exception.

We could copy the entry set to a local variable, but wouldn't we be back where we were before the patch?

Patch-set: 5