Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

In that case, the settings must be transmitted to the kernel on boot, otherwise you will end with kernel and settings not in sync. By the way, in that case the settings doesn't need to be stored as a global setting. it can be probably set internallly in the settings app preferences and restore in the bootreceiver.
Any other access to acccess the hardware state should be done via cmhardware interfaces (is there for a reason)

Patch-set: 2
Label: Code-Review=-1