Update patch set 1
Patch Set 1:

I know that there is a long-press-power-reboot option which just reboots the phone. What are the chances that when someone finds your phone on the street or steals it he actually knows how to enter recovery? My point is that this would give more time for damage control (or actually finding your device). The first thing thieves do when they get their hands on the phone is to either pull the battery out or shut down the phone and once that happens, you are stuck.

Patch-set: 1