Update patch set 2
Patch Set 2:

There's plenty of times when I have a lot of apps open that I'm not using and killing them individually is less convenient than just killing all of them.  Either way, it's useful for plenty of people and if it's not for you, it doesn't matter, since there's no disadvantage: it's an icon you don't have to use.

Patch-set: 2