Update patch set 1
Patch Set 1: Code-Review+1

This change was wrong in that it prevents recalculation of the ambient lux values after turning the screen on, effectively prolonging reaction time if the brightness conditions after screen on differs much from the conditions when the screen was previously turned off, as it disables the special handling at the beginning of the updateAmbientLux(long) method.

What is the exact problem this patch attempted to solve?

Patch-set: 1