Update patch set 1
Patch Set 1:

Yes, I think mine is better.  It just tells the compiler that it has to be careful when optimizing that variable and doesn't otherwise change the logic of the code.  Making it static changes the logic if this is ever called more than once concurrently.

Patch-set: 1
Label: Code-Review=0