Update patch set 1
Patch Set 1:

> I can write up a longer reply on why I think this is a bad idea
 > later but -1 from me.

I made this commit when M was first released to get gapps to work. It was the only way to get Setup Wizard to work and allow sign in. All it does is turn on the permissions for the core apps. If there was a better way I would have used it. Just thought I'd say why I made it.

Patch-set: 1