Update patch set 1
Patch Set 1:

This shouldn't affect themes other than it's another style they need to theme.  Is there a reason we want this while AOSP has removed it for the time being?

I see the original patch was addressing a bug so are we going to end up encountering that bug by re-enabling this?

Patch-set: 1