Update patch set 1
Patch Set 1:

Sorry, but I'm not sure I understand why passing a NULL pointer is needed.

To me, the problem appears to be that we are using a NULL pointer to signify two different things: a legal value equal to zero and "unsupported".

As I can look for a NULL pointer where I know that "NULL pointer == 0", one could look for a zero value where "NULL pointer == unsupported".

The thing is that doing the former is IMO strange, because I'd need a separate code path to treat a case that is not different from the others.

Either way, we are going to need to update several HALs (all the HAL changed in https://github.com/CyanogenMod/android_device_qcom_common/commit/d9668ba507d1c262e85c8ea58c4f1d43622c7a21 need a NULL pointer check). I simply thought it was better to change the code that is now looking for a NULL pointer and make it look for a zero value instead, but I might misunderstood something here and this is completely wrong.

Patch-set: 1